requirements.txt
Flask==2.3.3
Werkzeug==2.3.7
Jinja2==3.1.3
MarkupSafe==2.1.5
itsdangerous==2.1.2
click==8.1.7
SQLAlchemy==2.0.30
flask-sqlalchemy==3.1.1
flask-wtf==1.2.1
python-dotenv==1.1.0
Pillow==10.2.0
Babel==2.14.0
Flask-Babel==4.0.0
gunicorn==22.0.0
Markdown==3.5.2
pytz==2023.3.post1
alembic==1.12.1
bcrypt==4.1.2
cryptography==42.0.5
greenlet==3.0.3
psycopg2-binary==2.9.9
mysqlclient==2.2.4
python-dateutil==2.8.2
redis==5.0.1
requests==2.31.0
urllib3==2.2.1 